+++ wl_draw.c
@@ -1129,7 +1129,7 @@ void CalcTics (void)
     if(!tics)
     {
         // wait until end of current tic
-        SDL_Delay(((lasttimecount + 1) * 100) / 7 - curtime);
+        SDL_Delay(((((unsigned int)lasttimecount + 1) * 100) / 7 - (unsigned int)curtime));
         tics = 1;
     }
